TranchePricing: Adding a base for TranchePricing
Basic tranche handling done, calculatePrice() is now passing the tokensSold
all the way to the final getCurrentTranche(uint) function.

The isSane() is a difficult case, since in the original MilestonePricing
contract which is used as a base for this TranchePricing, "sanefulness" is
measured by comparing our start and end times to the ones in Crowdsale.
However, since we don't have any information to compare with the Crowdsale
contract, returning always "true" for now.