PreICOProxyBuyer: Inheriting Destructible as the escape hatch
If, for some reason, we can't get ethers out from the contract, this works as
the last resort. Before destroying the contract, one must first halt() the
contract, wait a while until people do not transfer money in anymore
(day or two), and after that destroy the contract. After destroying, contract
can't be used anymore.