KYCAttributes: Introducing enums for KYC attributes
When it comes to fixed numerical values, enums improve a lot the reading
experience of the code. And since KYC attribute is a number between 0-255,
enums fit here just fine. This was pointed out by @voith.

If we would have a base contract for KYC, the enum should be defined there.
But since we don't have one at the moment (but maybe in the future), I created
a new contract "KYCAttributes" which should be inherited by contracts needing
access to "KYCAttribute", the enum definition in question.

At first inheriting a contract to access enums might seem like a bad practice,
but my rationale is:

  - Enums are not exposed to the ABI anyway
  - Avoiding unnecessary redundancy when referencing an enum is a way to achieve
    readable code