CheckpointToken: Verify transactions only for transferFrom() and transfer()
Previously transactions were verified in transferInternal() which is not the
right place because it's used for functions like issueTokens(), burnTokens()
and forceTransfer(), which are intentionally with unlimited transfer power, and
hence should not go through a transaction verifier.

From now on, only transactions from transferFrom() and transfer() are verified.
Also, per EIP20 zero value  transactions are now permitted, and
transactionAgent should revert() instead of value=0.

Also version number was bumped to 0.3 because this change fundamentally changes
the operation of issueTokens(), burnTokens() and forceTransfer().